third_party/WebKit/LayoutTests/svg/animations/svgstring-animation-1-expected.txt

CONSOLE WARNING: SVG's SMIL animations (<animate>, <set>, etc.) are deprecated and will be removed. Please use CSS animations or Web animations instead.
SVG 1.1 dynamic animation tests

Test
Test animVal support for SVGAnimatedString animations.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS aElement.target.animVal is "base"
PASS aElement.target.baseVal is "base"
PASS aElement.target.animVal is "a"
PASS aElement.target.baseVal is "base"
PASS aElement.target.animVal is "a"
PASS aElement.target.baseVal is "base"
PASS aElement.target.animVal is "b"
PASS aElement.target.baseVal is "base"
PASS aElement.target.animVal is "b"
PASS aElement.target.baseVal is "base"
PASS aElement.target.animVal is "base"
PASS aElement.target.baseVal is "base"
PASS successfullyParsed is true

TEST COMPLETE
